Split form for coordinates
